feat: set alias in file content instead of link

+++ etc/nodenv.d/install/install-alias.bash
@@ -2,12 +2,12 @@
 
 resolve_versions() {
   if nodenv commands | grep "update-version-defs" 1>/dev/null 2>&1; then
-    NODE_BUILD_DEFINITIONS="$(nodenv-aliases --plugin_definitions_path):$NODE_BUILD_DEFINITIONS"
+    NODE_BUILD_DEFINITIONS="$NODE_BUILD_DEFINITIONS:$(nodenv aliases --plugin_definitions_path)"
     export NODE_BUILD_DEFINITIONS
   fi
 
-  if [ -L "$(nodenv-aliases --aliases_path)/$DEFINITION" ]; then
-    VERSION_NAME=$(nodenv-aliases --resolve_definition "$DEFINITION")
+  if [ -f "$(nodenv aliases --aliases_path)/$DEFINITION" ]; then
+    VERSION_NAME=$(nodenv aliases --resolve_definition "$DEFINITION")
     export VERSION_NAME
     export PREFIX="$NODENV_ROOT/versions/$VERSION_NAME"
     export DEFINITION="$VERSION_NAME"
@@ -32,12 +32,12 @@ resolve_versions() {
 
 alias_versions() {
   if [ "$STATUS" -eq 0 ] && [ -n "$VERSION_NAME" ]; then
-    if [ -f "$(nodenv-aliases --definitions_path)/$VERSION_NAME" ] || [ -f "$(nodenv-aliases --plugin_definitions_path)/$VERSION_NAME" ]; then
+    if [ -f "$(nodenv aliases --definitions_path)/$VERSION_NAME" ] || [ -f "$(nodenv aliases --plugin_definitions_path)/$VERSION_NAME" ]; then
       local ALIASES ALIAS
-      IFS=" " read -ra ALIASES <<< "$(nodenv-aliases --resolve_aliases "$VERSION_NAME")"
+      IFS=" " read -ra ALIASES <<< "$(nodenv aliases --resolve_aliases "$VERSION_NAME")"
 
       for ALIAS in "${ALIASES[@]}"; do
-        if [ -n "$ALIAS" ]; then
+        if [ -n "$ALIAS" ] && [ ! -L "$NODENV_ROOT/versions/$ALIAS" ]; then
           mkdir -p "$(dirname "$NODENV_ROOT/versions/$ALIAS")"
           ln -sf "${PREFIX:-$NODENV_ROOT/versions/$VERSION_NAME}" "$NODENV_ROOT/versions/$ALIAS"
           echo "Alias: $ALIAS -> $VERSION_NAME"
